Move empty-variations-image to the images folder so it can be used in other components
