+++ packages/js/product-editor/src/images/empty-variations-image.tsx
